Divide classes into packages
Create botto, botto.command, botto.util and botto.task packages
