+++ src/main/java/botto/DukeException.java
@@ -1,3 +1,5 @@
+package botto;
+
 public class DukeException extends Exception{
     public DukeException(String message) {
         super(message);
