+++ src/main/java/botto/command/AddEventCommand.java
@@ -1,3 +1,12 @@
+package botto.command;
+
+import botto.*;
+import botto.task.Event;
+import botto.task.Task;
+import botto.util.Storage;
+import botto.util.TaskList;
+import botto.util.Ui;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
