+++ src/main/java/botto/task/Event.java
@@ -1,3 +1,5 @@
+package botto.task;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
