+++ src/main/java/botto/task/Task.java
@@ -1,3 +1,5 @@
+package botto.task;
+
 public class Task {
     protected String description;
     protected boolean isDone;
