+++ src/main/java/botto/task/Todo.java
@@ -1,3 +1,5 @@
+package botto.task;
+
 public class Todo extends Task{
 
     public Todo(String description) {
