+++ src/main/java/botto/util/Ui.java
@@ -1,8 +1,12 @@
+package botto.util;
+
+import botto.task.Task;
+
 import java.util.List;
 import java.util.Scanner;
 
 public class Ui {
-    private final String BOT_NAME = "Botto";
+    private final String BOT_NAME = "botto.Botto";
     private final String INDENTATION = "    ";
 
     private Scanner scanner;
